backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
python-jose[cryptography]
passlib[bcrypt]